perf(gang): lift weapon table row derivation out of the render path
The weapon table rebuilt its entire grouping/sorting/trait pipeline in the
component body on every render. In a CPU profile of a 76-fighter gang page it
was the hottest application code: the per-profile composite map key alone was
10.3% of self time, with the surrounding grouping frames adding ~6% more.

Move the pipeline to utils/weapon-variants.ts as a pure function and call it
from a single useMemo keyed on the weapon set, then fix the two things the
profile pointed at:

- The map key interpolated weaponProfileSignature -- the joined signature of
  every base profile on the weapon -- once per profile. Only the group id
  varies per profile, so the rest is now interned to a small integer per
  weapon. Group ids are UUIDs and contain no "|", so the partition is
  unchanged.
- localeCompare(x, undefined, {sensitivity:'base'}) constructs an Intl.Collator
  per comparison; measured 46x slower than a cached instance. Three sites now
  share two module-level collators -- two instances, because the traits sort
  used bare localeCompare() (default sensitivity) and merging them would
  reorder traits. The traits sort was the highest-frequency site: once per
  rendered row.

Also hoists the formatters object and formatStrength to module scope; they
close over nothing and were re-allocated per card per render.

Behaviour is unchanged. Verified by rendering 458 cases -- 25 weapon fixtures
covering duplicate counts, master-crafted detection, ammo/special profiles,
weapon_group_id grouping, hardpoint arcs, orphan specials and numeric profile
names, crossed with entity/edition/view-mode, plus full fighter lists at 7/12/
20/76 fighters -- through renderToString before and after: byte-identical
output. Isolated, the 76 weapon tables render 24.9ms -> 16.1ms; in the profile
the weapon pipeline drops from ~17% of app frames to ~3.5%.
